Restore old behavior in FrameInstance#getCallNode()
